\name{DEFAULT_CIRC_SEQS}
\alias{DEFAULT_CIRC_SEQS}
\docType{data}
\title{character vector: strings that are usually circular chromosomes}
\usage{DEFAULT_CIRC_SEQS}
\description{
The \code{DEFAULT_CIRC_SEQS} character vector contains strings that are
normally used by major repositories as the names of chromosomes that
are typically circular, it is available as a convenience so that users
can us it as a default value for \code{circ_seqs} arguments, and
append to it as needed.
}
\seealso{
  \code{\link{makeTxDbFromUCSC}},
  \code{\link{makeTxDbFromBiomart}}
}

\examples{
DEFAULT_CIRC_SEQS
}
\keyword{datasets}
