\name{DEFAULTCIRCSTRS}
\alias{DEFAULTCIRCSTRS}
\docType{data}
\title{character vector: strings that are usually circular chromosomes}
\usage{DEFAULTCIRCSTRS}
\description{
The \code{DEFAULTCIRCSTRS} character vector contains strings that are
normally used by major repositories as the names of chromosomes that
are typically circular, it is available as a convenience so that users
can us it as a default value for \code{circ_seqs} arguments, and
append to it as needed.
}
\examples{
DEFAULTCIRCSTRS
}
\keyword{datasets}
