\name{GenomicFeatures-defunct}
\alias{GenomicFeatures-defunct}
\title{Defunct Functions in Package \pkg{GenomicFeatures}}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{transcripts_deprecated}
\alias{exons_deprecated}
\alias{introns_deprecated}
%
\description{
  The functions or variables listed here are no longer part of the GenomicFeatures package.
}
\usage{
transcripts_deprecated()
exons_deprecated()
introns_deprecated()
}
\seealso{
  \code{\link{Defunct}}
}

