\name{GenomicFeatures-deprecated}
\alias{GenomicFeatures-deprecated}
\title{Deprecated Functions in package \pkg{GenomicFeatures}}
%------ PLEASE: put \alias{.} here for  EACH !
%\alias{isActiveSeq}
%\alias{isActiveSeq<-}
%\alias{isActiveSeq,TranscriptDb-method}
%\alias{isActiveSeq<-,TranscriptDb-method}
%
\description{
  The functions or variables listed here have been deprecated and should no
  longer be used.
}
\seealso{
  \code{\link{Deprecated}}
}

