\name{makeFeatureDbFromUCSC}

\alias{supportedUCSCFeatureDbTracks}
\alias{supportedUCSCFeatureDbTables} 
\alias{UCSCFeatureDbTableSchema}
\alias{makeFeatureDbFromUCSC}

\title{
  [Moved to txdbmaker] Make a FeatureDb object from annotations available
  at the UCSC Genome Browser
}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, functions
  \code{supportedUCSCFeatureDbTracks()}, \code{supportedUCSCFeatureDbTables()},
  \code{UCSCFeatureDbTableSchema()}, and \code{makeFeatureDbFromUCSC()} are
  defined in the \pkg{txdbmaker} package.
}

\seealso{
  \code{txdbmaker::\link[txdbmaker]{supportedUCSCFeatureDbTracks}},
  \code{txdbmaker::\link[txdbmaker]{supportedUCSCFeatureDbTables}},
  \code{txdbmaker::\link[txdbmaker]{UCSCFeatureDbTableSchema}},
  and \code{txdbmaker::\link[txdbmaker]{makeFeatureDbFromUCSC}}
  in the \pkg{txdbmaker} package.
}

