\name{makeTxDb}

\alias{makeTxDb}

\title{
  [Moved to txdbmaker] Make a TxDb object from user supplied annotations
}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, the \code{makeTxDb} function
  is defined in the \pkg{txdbmaker} package.
}

\seealso{
  \code{txdbmaker::\link[txdbmaker]{makeTxDb}} in the \pkg{txdbmaker}
  package.
}

