\name{makeTxDbFromBiomart}

\alias{makeTxDbFromBiomart}
\alias{getChromInfoFromBiomart}

\title{
  [Moved to txdbmaker] Make a TxDb object from annotations available
  on a BioMart database
}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, functions
  \code{makeTxDbFromBiomart()} and \code{getChromInfoFromBiomart()}
  are defined in the \pkg{txdbmaker} package.
}

\seealso{
  \code{txdbmaker::\link[txdbmaker]{makeTxDbFromBiomart}} and
  \code{txdbmaker::\link[txdbmaker]{getChromInfoFromBiomart}} in
  the \pkg{txdbmaker} package.
}

