\name{makeTxDbFromEnsembl}

\alias{makeTxDbFromEnsembl}

\title{
  [Moved to txdbmaker] Make a TxDb object from an Ensembl database
}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, the \code{makeTxDbFromEnsembl}
  function is defined in the \pkg{txdbmaker} package.
}

\seealso{
  \code{txdbmaker::\link[txdbmaker]{makeTxDbFromEnsembl}} in the
  \pkg{txdbmaker} package.
}

