\name{makeTxDbFromGFF}

\alias{makeTxDbFromGFF}

\title{
  [Moved to txdbmaker] Make a TxDb object from annotations available
  as a GFF3 or GTF file
}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, the \code{makeTxDbFromGFF}
  function is defined in the \pkg{txdbmaker} package.
}

\seealso{
  \code{txdbmaker::\link[txdbmaker]{makeTxDbFromGFF}} in the \pkg{txdbmaker}
  package.
}

