\name{makeTxDbFromGRanges}

\alias{makeTxDbFromGRanges}

\title{[Moved to txdbmaker] Make a TxDb object from a GRanges object}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, the \code{makeTxDbFromGRanges}
  function is defined in the \pkg{txdbmaker} package.
}

\seealso{
  \code{txdbmaker::\link[txdbmaker]{makeTxDbFromGRanges}} in the
  \pkg{txdbmaker} package.
}

