\name{makeTxDbFromUCSC}

\alias{supportedUCSCtables}
\alias{browseUCSCtrack}
\alias{makeTxDbFromUCSC}

\title{
  [Moved to txdbmaker] Make a TxDb object from annotations available
  at the UCSC Genome Browser
}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, functions
  \code{makeTxDbFromUCSC()}, \code{supportedUCSCtables()},
  and \code{browseUCSCtrack()} are defined in the \pkg{txdbmaker}
  package.
}

\seealso{
  \code{txdbmaker::\link[txdbmaker]{makeTxDbFromUCSC}},
  \code{txdbmaker::\link[txdbmaker]{supportedUCSCtables}},
  and \code{txdbmaker::\link[txdbmaker]{browseUCSCtrack}}
  in the \pkg{txdbmaker} package.
}

