\name{makeTxDbPackage}

\alias{supportedMiRBaseBuildValues}
\alias{makePackageName}
\alias{makeTxDbPackage}
\alias{makeTxDbPackageFromUCSC}
\alias{makeFDbPackageFromUCSC}
\alias{makeTxDbPackageFromBiomart}

\title{
  [Moved to txdbmaker] Making a TxDb package from annotations available
  at the UCSC Genome Browser, biomaRt or from another source.
}

\description{
  IMPORTANT NOTE: Starting with BioC 3.19, functions
  \code{makeTxDbPackageFromUCSC()}, \code{makeFDbPackageFromUCSC()},
  \code{makeTxDbPackageFromBiomart()}, \code{makeTxDbPackage()}
  \code{supportedMiRBaseBuildValues()} and \code{makePackageName()}
  are defined in the \pkg{txdbmaker} package.
}

\seealso{
  \code{txdbmaker::\link[txdbmaker]{makeTxDbPackageFromUCSC}},
  \code{txdbmaker::\link[txdbmaker]{makeFDbPackageFromUCSC}},
  \code{txdbmaker::\link[txdbmaker]{makeTxDbPackageFromBiomart}},
  \code{txdbmaker::\link[txdbmaker]{makeTxDbPackage}},
  \code{txdbmaker::\link[txdbmaker]{supportedMiRBaseBuildValues}},
  and \code{txdbmaker::\link[txdbmaker]{makePackageName}}
  in the \pkg{txdbmaker} package.
}

