% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microRNAs.R
\name{microRNAs}
\alias{microRNAs}
\alias{microRNAs,TxDb-method}
\alias{tRNAs}
\alias{tRNAs,TxDb-method}
\title{Extract microRNA or tRNA genomic ranges from an object}
\usage{
microRNAs(x)
}
\arguments{
\item{x}{A \link{TxDb} object.}
}
\value{
A \link[GenomicRanges]{GRanges} object.
}
\description{
Generic functions to extract microRNA or tRNA genomic ranges from an object.
This page documents the methods for \link{TxDb} objects only.
}
\examples{

\dontrun{library(TxDb.Hsapiens.UCSC.hg19.knownGene)
library(mirbase.db)
microRNAs(TxDb.Hsapiens.UCSC.hg19.knownGene)
}

}
\seealso{
\itemize{ \item \code{\link{transcripts}},
\code{\link{transcriptsBy}}, and \code{\link{transcriptsByOverlaps}} for the
core genomic features extractors.  \item The \link{TxDb} class.  }
}
\author{
M. Carlson
}
\keyword{methods}
