\name{GenomicFeatures-deprecated}
\alias{GenomicFeatures-deprecated}
\title{Deprecated Functions in package \pkg{GenomicFeatures}}
%------ PLEASE: put \alias{.} here for  EACH !
\alias{determineDefaultSeqnameStyle}
\alias{determineDefaultSeqnameStyle,TranscriptDb-method}
%
\description{
  The functions or variables listed here have been deprecated and should no
  longer be used.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{determineDefaultSeqnameStyle: \code{\link{seqlevelsStyle}}}

  }
}

\seealso{
  \code{\link{Deprecated}}
}

